Implement python-server command for enhanced Python app deployment with ASGI/WSGI support and improved CLI options
