+++ cmd/cli/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "caddysnake"
-version = "0.1.10"
+version = "0.2.0"
 description = "Python WSGI/ASGI server powered by Caddy"
 authors = [
     {name = "Miguel Liezun", email = "liezun.js@gmail.com"},
@@ -23,11 +23,13 @@ classifiers = [
     "Programming Language :: Python :: 3.13",
     "Programming Language :: Go",
 ]
+dependencies = [
+    "click>=8.3.0",
+]
 
 [project.urls]
-Homepage = "https://github.com/mliezun/caddy-snake"
+Homepage = "https://caddy-snake.readthedocs.io"
 Repository = "https://github.com/mliezun/caddy-snake.git"
-# Documentation = "https://caddy-snake.readthedocs.io/"
 
 [tool.setuptools]
 packages = ["caddysnake"]
